refactor: 獲得トースト演出をフック化
